[fix] all routes of users handle, create a user into the table when connected for the first time handle
