+++ backend/src/routes/users/users.query.ts
@@ -3,6 +3,9 @@
  * Licensed under the MIT License.
  */
 
+import { dataBase } from '../../config/db.js';
+import { User } from '../../entity/user.js';
+
 async function userConnection(endpoint: string, accessToken: string): Promise<any> {
     const options = {
         headers: {
@@ -20,4 +23,15 @@ async function userConnection(endpoint: string, accessToken: string): Promise<an
     }
 }
 
-export { userConnection };
+async function createUser(userData: any): Promise<any> {
+    let exist = await dataBase.getRepository(User).createQueryBuilder('user').where('user.email = :email', { email: userData.userPrincipalName }).getOne();
+    if (!exist) {
+        const user = new User();
+        user.email = userData.mail;
+        user.token = userData.id;
+        user.role = false;
+        const data = await dataBase.getRepository(User).save(user);
+    }
+};
+
+export { userConnection, createUser };
